+++ lib/swagcov/errors.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Swagcov
+  module Errors
+    class BadConfiguration < ::RuntimeError
+    end
+  end
+end