+++ lib/swagcov/openapi_files.rb
@@ -9,7 +9,7 @@ module Swagcov
 
     def find_response_keys path:, route_verb:
       # replace :id with {id}
-      regex = Regexp.new("^#{path.gsub(%r{:[^/]+}, '\\{[^/]+\\}')}?$")
+      regex = ::Regexp.new("^#{path.gsub(%r{:[^/]+}, '\\{[^/]+\\}')}?$")
 
       matching_paths_key = @openapi_paths.keys.grep(regex).first
       matching_request_method_key = @openapi_paths.dig(matching_paths_key, route_verb.downcase)
@@ -20,15 +20,15 @@ module Swagcov
     private
 
     def load_yamls
-      Dir.glob(@filepaths).reduce({}) do |hash, filepath|
+      ::Dir.glob(@filepaths).reduce({}) do |hash, filepath|
         hash.merge(load_yaml(filepath))
       end
     end
 
     def load_yaml filepath
-      YAML.load_file(filepath)["paths"]
-    rescue Psych::SyntaxError
-      raise BadConfigurationError, "Malinformed openapi file (#{filepath})"
+      ::YAML.load_file(filepath)["paths"]
+    rescue ::Psych::SyntaxError
+      raise ::Swagcov::Errors::BadConfiguration, "Malinformed openapi file (#{filepath})"
     end
   end
 end